hosting: fix issue with Apache mod_proxy_fcgi settings on Ubuntu 22.04
Under Apache 2.4.52, if a `ProxyPassMatch` worker overlaps with a
`Proxy` worker, neither seem to be configured, although this may be
limited to the fcgi-over-socket use case (the error Apache reports is a
DNS lookup failure on the `fcgi://` hostname, which is arbitrary in
this scenario).

Resolved by replacing `ProxyPassMatch` with `SetHandler proxy:`,
ensuring only one proxy worker is configured per PHP-FPM site.